Revert "fix(condo): DOMA-11788 create common canManageEntityWithOrganization function"
This reverts commit a5f940eadcdc7cf244d2c785adebb09cc32a9442.
