+++ apps/condo/domains/meter/access/Meter.js
@@ -1,7 +1,12 @@
 /**
  * Generated by `createschema meter.Meter 'number:Text; billingAccountMeter?:Relationship:BillingAccountMeter:SET_NULL; account?:Relationship:BillingAccount:SET_NULL; property:Relationship:Property:CASCADE; unitName:Text; place?:Text; resource:Relationship:MeterResource:CASCADE;'`
  */
+
+const { get } = require('lodash')
+const uniq = require('lodash/uniq')
+
 const { throwAuthenticationError } = require('@open-condo/keystone/apolloErrorFormatter')
+const { find } = require('@open-condo/keystone/schema')
 
 const { getAvailableResidentMeters } = require('@condo/domains/meter/utils/serverSchema')
 const {
@@ -9,8 +14,8 @@ const {
     canManageObjectsAsB2BAppServiceUser,
 } = require('@condo/domains/miniapp/utils/b2bAppServiceUserAccess')
 const {
+    checkPermissionsInEmployedOrRelatedOrganizations,
     getEmployedOrRelatedOrganizationsByPermissions,
-    canManageEntityWithOrganization,
 } = require('@condo/domains/organization/utils/accessSchema')
 const { RESIDENT, SERVICE } = require('@condo/domains/user/constants/common')
 
@@ -47,7 +52,7 @@ async function canReadMeters (args) {
 }
 
 async function canManageMeters (args) {
-    const { authentication: { item: user } } = args
+    const { authentication: { item: user }, originalInput, operation, itemId, itemIds, context } = args
 
     if (!user) return throwAuthenticationError()
     if (user.deletedAt) return false
@@ -57,7 +62,33 @@ async function canManageMeters (args) {
         return await canManageObjectsAsB2BAppServiceUser(args)
     }
 
-    return await canManageEntityWithOrganization(args, 'canManageMeters', true)
+    const isBulkRequest = Array.isArray(originalInput)
+    let organizationIds
+
+    if (operation === 'create') {
+        if (isBulkRequest) {
+            organizationIds = originalInput.map(el => get(el, ['data', 'organization', 'connect', 'id']))
+
+            if (organizationIds.filter(Boolean).length !== originalInput.length) return false
+            organizationIds = uniq(organizationIds)
+        } else {
+            const organizationId = get(originalInput, ['organization', 'connect', 'id'])
+            if (!organizationId) return false
+            organizationIds = [organizationId]
+        }
+    } else if (operation === 'update') {
+        const ids = itemIds || [itemId]
+        if (ids.length !== uniq(ids).length) return false
+
+        const items = await find('Meter', {
+            id_in: ids,
+            deletedAt: null,
+        })
+        if (items.length !== ids.length || items.some(item => !item.organization)) return false
+        organizationIds = uniq(items.map(item => item.organization))
+    }
+
+    return await checkPermissionsInEmployedOrRelatedOrganizations(context, user, organizationIds, 'canManageMeters')
 }
 
 /*
