+++ apps/condo/domains/meter/access/PropertyMeter.js
@@ -1,11 +1,16 @@
 /**
  * Generated by `createschema meter.PropertyMeter 'number:Text; numberOfTariffs:Integer; installationDate:DateTimeUtc; commissioningDate:DateTimeUtc; verificationDate:DateTimeUtc; nextVerificationDate:DateTimeUtc; controlReadingsDate:DateTimeUtc; sealingDate:DateTimeUtc; isAutomatic:Checkbox; organization:Relationship:Organization:CASCADE; property:Relationship:Property:CASCADE; resource:Relationship:MeterResource:CASCADE; meta:Json;'`
  */
+
+const get = require('lodash/get')
+const uniq = require('lodash/uniq')
+
 const { throwAuthenticationError } = require('@open-condo/keystone/apolloErrorFormatter')
+const { find } = require('@open-condo/keystone/schema')
 
 const {
+    checkPermissionsInEmployedOrRelatedOrganizations,
     getEmployedOrRelatedOrganizationsByPermissions,
-    canManageEntityWithOrganization,
 } = require('@condo/domains/organization/utils/accessSchema')
 
 async function canReadPropertyMeters ({ authentication: { item: user }, context }) {
@@ -23,14 +28,38 @@ async function canReadPropertyMeters ({ authentication: { item: user }, context 
     }
 }
 
-async function canManagePropertyMeters (args) {
-    const { authentication: { item: user } } = args
-
+async function canManagePropertyMeters ({ authentication: { item: user }, originalInput, operation, itemId, itemIds, context }) {
     if (!user) return throwAuthenticationError()
     if (user.deletedAt) return false
     if (user.isAdmin) return true
 
-    return await canManageEntityWithOrganization(args, 'canManageMeters', true)
+    const isBulkRequest = Array.isArray(originalInput)
+    let organizationIds
+
+    if (operation === 'create') {
+        if (isBulkRequest) {
+            organizationIds = originalInput.map(el => get(el, ['data', 'organization', 'connect', 'id']))
+
+            if (organizationIds.filter(Boolean).length !== originalInput.length) return false
+            organizationIds = uniq(organizationIds)
+        } else {
+            const organizationId = get(originalInput, ['organization', 'connect', 'id'])
+            if (!organizationId) return false
+            organizationIds = [organizationId]
+        }
+    } else if (operation === 'update') {
+        const ids = itemIds || [itemId]
+        if (ids.length !== uniq(ids).length) return false
+
+        const items = await find('PropertyMeter', {
+            id_in: ids,
+            deletedAt: null,
+        })
+        if (items.length !== ids.length || items.some(item => !item.organization)) return false
+        organizationIds = uniq(items.map(item => item.organization))
+    }
+
+    return await checkPermissionsInEmployedOrRelatedOrganizations(context, user, organizationIds, 'canManageMeters')
 }
 
 /*
