+++ apps/condo/domains/meter/utils/testSchema/index.js
@@ -345,8 +345,8 @@ async function updateTestPropertyMeter (client, id, extraAttrs = {}) {
         sender,
         ...extraAttrs,
     }
-    const obj = await PropertyMeter.update(client, id, attrs)
-    return [obj, attrs]
+
+    return PropertyMeter.update(client, id, attrs)
 }
 
 async function updateTestPropertyMeters (client, updateAttrs = []) {
